removes card directory and moves Cards to Components directory, and fixes imports
