+++ src/components/CardBlank/CardBlank.scss
