+++ src/components/CardEvent/CardEvent.scss
