+++ src/components/CardLarge/CardLarge.scss
