+++ src/components/CardSmall/CardSmall.js
