+++ src/components/CardSmall/CardSmall.scss
