+++ src/components/components.scss
@@ -1,10 +1,10 @@
 @import
   'Button/Button',
-  'Cards/CardLarge/CardLarge',
-  'Cards/CardSmall/CardSmall',
-  'Cards/CardSquare/CardSquare',
-  'Cards/CardBlank/CardBlank',
-  'Cards/CardEvent/CardEvent',
+  'CardLarge/CardLarge',
+  'CardSmall/CardSmall',
+  'CardSquare/CardSquare',
+  'CardBlank/CardBlank',
+  'CardEvent/CardEvent',
   'Container/Container',
   'Footer/Footer',
   'Layout/Layout',
